+++ Source/EventFlow.Tests/UnitTests/Sagas/AggregateSagas/SagaAggregateStoreTests.cs
@@ -23,6 +23,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
@@ -30,6 +31,7 @@ using EventFlow.Core;
 using EventFlow.Core.Caching;
 using EventFlow.Sagas.AggregateSagas;
 using EventFlow.TestHelpers;
+using EventFlow.TestHelpers.Aggregates.Events;
 using EventFlow.TestHelpers.Aggregates.Sagas;
 using Moq;
 using NUnit.Framework;
@@ -39,9 +41,15 @@ namespace EventFlow.Tests.UnitTests.Sagas.AggregateSagas
     [Category(Categories.Unit)]
     public class SagaAggregateStoreTests : TestsFor<SagaAggregateStore>
     {
+        private ThingySagaId _thingySagaId;
+        private Mock<ThingySaga> _thingySaga;
+
         [SetUp]
         public void SetUp()
         {
+            _thingySagaId = A<ThingySagaId>();
+            _thingySaga = InjectMock<ThingySaga>(_thingySagaId);
+
             Inject<IMemoryCache>(A<MemoryCache>());
         }
 
@@ -49,16 +57,9 @@ namespace EventFlow.Tests.UnitTests.Sagas.AggregateSagas
         public async Task AggregateStore_UpdateAsync_IsInvoked()
         {
             // Arrange
-            var aggregateStoreMock = InjectMock<IAggregateStore>();
             var thingySagaId = A<ThingySagaId>();
             var sourceId = A<SourceId>();
-            aggregateStoreMock
-                .Setup(s => s.UpdateAsync(
-                    thingySagaId,
-                    sourceId,
-                    It.IsAny<Func<ThingySaga, CancellationToken, Task>>(),
-                    It.IsAny<CancellationToken>()))
-                .ReturnsAsync(new List<IDomainEvent>());
+            var aggregateStoreMock = ArrangeAggregateStoreMock(thingySagaId, sourceId, A<bool>());
 
             // Act
             await Sut.UpdateAsync(
@@ -76,6 +77,51 @@ namespace EventFlow.Tests.UnitTests.Sagas.AggregateSagas
                     It.IsAny<Func<ThingySaga, CancellationToken, Task>>(),
                     It.IsAny<CancellationToken>()),
                 Times.Once);
+        }
+
+        [TestCase(true)]
+        [TestCase(false)]
+        public async Task CommandBus_PublishAsync_IsInvoked(bool eventsAreEmitted)
+        {
+            // Arrange
+            var thingySagaId = A<ThingySagaId>();
+            var sourceId = A<SourceId>();
+            ArrangeAggregateStoreMock(thingySagaId, sourceId, eventsAreEmitted);
+
+            // Act
+            await Sut.UpdateAsync(
+                thingySagaId,
+                typeof(ThingySaga),
+                sourceId,
+                (s, c) => Task.FromResult(0),
+                CancellationToken.None);
+
+            // Assert
+            _thingySaga.Verify(
+                s => s.PublishAsync(It.IsAny<ICommandBus>(), It.IsAny<CancellationToken>()),
+                Times.AtLeastOnce);
+        }
+
+        private Mock<IAggregateStore> ArrangeAggregateStoreMock(
+            ThingySagaId thingySagaId,
+            ISourceId sourceId,
+            bool returnsDomainEvents)
+        {
+            var aggregateStoreMock = InjectMock<IAggregateStore>();
+            var domainEvents = ManyDomainEvents<ThingyPingEvent>(returnsDomainEvents ? 3 : 0)
+                .Cast<IDomainEvent>()
+                .ToList();
+
+            aggregateStoreMock
+                .Setup(s => s.UpdateAsync(
+                    thingySagaId,
+                    sourceId,
+                    It.IsAny<Func<ThingySaga, CancellationToken, Task>>(),
+                    It.IsAny<CancellationToken>()))
+                .Callback<ThingySagaId, ISourceId, Func<ThingySaga, CancellationToken, Task>, CancellationToken>(
+                    (id, sid, f, c) => f(_thingySaga.Object, CancellationToken.None).Wait(c))
+                .ReturnsAsync(domainEvents);
+            return aggregateStoreMock;
         }
     }
 }